okx_trading_bot/requirements.txt

# Core trading and API libraries
ccxt>=4.0.0
okx>=2.1.0
requests>=2.31.0
websocket-client>=1.7.0

# Ollama AI integration (free AI)
ollama>=0.1.0

# Technical analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Discord integration
discord.py>=2.3.0
discord-webhook>=1.3.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# PDF reporting
reportlab>=4.0.0
matplotlib>=3.7.0
plotly>=5.15.0
kaleido>=0.2.1

# Database and data storage
aiosqlite>=0.19.0

# Async and concurrency
aiohttp>=3.8.0

# Security and encryption
cryptography>=40.0.0

# Machine learning for AI assistant
scikit-learn>=1.3.0

# Utilities
schedule>=1.2.0
pytz>=2023.3
python-dateutil>=2.8.0
tqdm>=4.65.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
